Decoupled Auth and Key from mixnet app
Moved Auth and Key to base because both models are used independetly by mixnet and voting
