+++ votinus/base/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
 
-# Register your models here.
+from .models import Auth, Key
+
+
+admin.site.register(Auth)
+admin.site.register(Key)
