+++ votinus/base/serializers.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from .models import Auth, Key
+
+
+class AuthSerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Auth
+        fields = ('name', 'url', 'me')
+
+
+class KeySerializer(serializers.HyperlinkedModelSerializer):
+    class Meta:
+        model = Key
+        fields = ('p', 'g', 'y')
