+++ votinus/mixnet/admin.py
@@ -1,8 +1,6 @@
 from django.contrib import admin
 
-from .models import Auth, Key, Mixnet
+from .models import Mixnet
 
 
-admin.site.register(Auth)
-admin.site.register(Key)
 admin.site.register(Mixnet)
