+++ votinus/mixnet/views.py
@@ -1,12 +1,12 @@
+from django.conf import settings
+from django.shortcuts import get_object_or_404
 from rest_framework import viewsets
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from .serializers import MixnetSerializer, KeySerializer, AuthSerializer
+from .serializers import MixnetSerializer
 from .models import Auth, Mixnet, Key
-
-from django.conf import settings
-from django.shortcuts import get_object_or_404
+from base.serializers import KeySerializer, AuthSerializer
 
 
 class MixnetViewSet(viewsets.ModelViewSet):
