+++ votinus/votinus/settings.py
@@ -42,7 +42,6 @@ INSTALLED_APPS = [
     'django_filters',
     'rest_framework',
     'rest_framework.authtoken',
-
 ]
 
 REST_FRAMEWORK = {
